// frontend/src/shared/lib/api/model/commentResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */
import type { CommentResponseText } from './commentResponseText';
import type { CommentResponsePayload } from './commentResponsePayload';
import type { CommentResponseEvidence } from './commentResponseEvidence';

export interface CommentResponse {
  comment_type: string;
  text?: CommentResponseText;
  payload?: CommentResponsePayload;
  evidence?: CommentResponseEvidence;
}